pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "gradio>=6.4.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "tiktoken>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]
